third_party/WebKit/LayoutTests/inspector/profiler/agents-disabled-check-expected.txt

Test that if a profiler is working all the agents are disabled.

--> WebInspector.targetManager.suspendAllTargets();
frontend: {"id":<number>,"method":"Worker.setAutoconnectToWorkers","params":{"value":false}}
frontend: {"id":<number>,"method":"Debugger.disable"}
frontend: {"id":<number>,"method":"Debugger.setAsyncCallStackDepth","params":{"maxDepth":0}}
frontend: {"id":<number>,"method":"Page.setOverlayMessage"}
frontend: {"id":<number>,"method":"DOM.disable"}
frontend: {"id":<number>,"method":"CSS.disable"}

--> WebInspector.targetManager.resumeAllTargets();
frontend: {"id":<number>,"method":"Worker.setAutoconnectToWorkers","params":{"value":true}}
frontend: {"id":<number>,"method":"Debugger.enable"}
frontend: {"id":<number>,"method":"Debugger.setPauseOnExceptions","params":{"state":"none"}}
frontend: {"id":<number>,"method":"Debugger.setAsyncCallStackDepth","params":{"maxDepth":4}}
frontend: {"id":<number>,"method":"DOM.enable"}
frontend: {"id":<number>,"method":"CSS.enable"}

--> done
